day 18

+++ app/Http/Controllers/ProjectController.php
@@ -11,18 +11,25 @@ class ProjectController extends Controller
 
     public function index()
     {
-        $projects = Project::with(['owner', 'assignee'])->get();
+
+        $this->authorize('viewAny', Project::class);
+
+        $projects = Project::with(['owner', 'assignee'])->orderBy('id', 'asc')->get();
 
         return view('pages.Projects.index', compact('projects'));
     }
 
     public function create()
     {
+        $this->authorize('create', Project::class);
+
         return view('pages.Projects.create');
     }
 
     public function store(StoreProjectRequest $request)
     {
+        $this->authorize('create', Project::class);
+
         Project::create($request->validated());
 
         return redirect()->route('projects.create', ['access' => 'yes'])->with('success', 'Проект создан');
@@ -32,6 +39,8 @@ class ProjectController extends Controller
     {
         $project = Project::with(['owner', 'assignee'])->findOrFail($id);
 
+        $this->authorize('view', $project);
+
         return view('pages.Projects.show', compact('project'));
     }
 
@@ -40,12 +49,16 @@ class ProjectController extends Controller
 
         $project = Project::findOrFail($id);
 
+        $this->authorize('view', $project);
+
         return view('pages.Projects.edit', compact('project'));
     }
 
     public function update(UpdateProjectRequest $request, $id)
     {
         $project = Project::findOrFail($id);
+
+        $this->authorize('update', $project);
 
         $project->update($request->validated());
 
@@ -55,9 +68,12 @@ class ProjectController extends Controller
     public function destroy($id)
     {
         $project = Project::findOrFail($id);
+
+        $this->authorize('delete', $project);
+
         $project->delete();
 
-        return redirect()->route('projects.index');
+        return redirect()->route('projects.index')->with('success', 'Проект удалён');
     }
 
 }
